always use included rfkill
This isn't really desirable, but for now fixes build on
some kernels.

What we really should do is rename all the rfkill to be
backport symbols, and use it only on kernels that didn't
have a usable rfkill. OTOH, on new kernels one may want
to use a backported rfkill if the base configuration
doesn't include it?

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>