backports: add rndis_wlan usbnet requirements
Apparently I missed this during the conversion to the new
copy-list etc. The rndis_wlan driver requires some usbnet
and rndis code, if that's not present then it is either
not selectable or can be compiled but won't link against
older kernels that don't have the right requirements (or
don't have the symbols exported.)

For now, because the makefile parser is stupid, patch out
the asix driver objects because otherwise it's possible
to select it and fail the build. Improving the parser is
possible but I need to finish other things first.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>