backport build system: make autoconf header depend on Kconfig
If the kernel's configuration changes for some reason or you
change the KLIB path then the Kconfig.kernel file is rebuilt
correctly, but the .config and autoconf header may be left
in inconsistent state. Make the autoconf header make target
depend on the Kconfig.kernel and Kconfig.versions file, this
fixes the issue since then oldconfig will run and the header
will be rebuilt.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>