backports: revert remove_proc_subtree() for backport
Intorduced in next-20130429 and Linus has merged it
onto v3.10-rc1. We cannot backport remove_proc_subtree()
modularly [0] so just go ahead and revert this change.

The alternative is to go and review each procfs usage
for each driver we have issues and see if it makes
sense to replace procs with debugfs upstream.

commit 8ce584c7416d8a85a6f3edc17d1cddefe331e87e
Author: Al Viro <viro@zeniv.linux.org.uk>
Date:   Sat Mar 30 20:13:46 2013 -0400

    procfs: add proc_remove_subtree()

    just what it sounds like; do that only to procfs subtrees you've
    created - doing that to something shared with another driver is
    not only antisocial, but might cause interesting races with
    proc_create() and its ilk.

    Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>

[0] http://marc.info/?t=136841961600002&r=1&w=2

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>