backports: fix firmware install for old udev on Ubuntu
Rules for older versions of udev (e.g. version 113-0ubuntu16 on Ubuntu 7.10) are
looked for in /etc/udev/rules.d. For later versions, they are looked for in
either the default /lib/udev/rules.d location or the custom /etc/udev/rules.d
location. For backwards compatibility with older udev, the rules should be
installed in /etc/udev/rules.d.

Signed-off-by: Anoop Karollil <anoop.karollil@gmail.com>
Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>