backports: add RedHat version
We have to deactivate some backports when a RedHat kernel was found,
the LINUX_BACKPORT preprocessor statement is not sufficient enough.
This is done based on RHEL_RELEASE_CODE. When backports is compiled
against a non RedHat kernel the RHEL_RELEASE_CODE is set to 0, so these
defines are defined to some value.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>
Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>