backports: no kfifo.h when !CPTCFG_BACKPORT_BUILD_KFIFO
When CPTCFG_BACKPORT_BUILD_KFIFO is not selected the default kernel
header file should be used and not the backport code from backports. On
RHEL 6.4 the kfido-new.h is included instead of the normal kfifo.h.

The kfifo backport code is only build when CPTCFG_BACKPORT_BUILD_KFIFO
is set and not depending on some kernel version.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>
Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>