backports: Fix and document EXPORT_SYMBOL_GPL() preference
In terms of project naming the backports project has evolved
as follows:

compat-wireless --> compat-drivers --> backports

All along the design and intent behind the initial project
has been to provide a framework for delivery of *upstream*
drivers using a backported infrastructure. I've made it
clear that in no way shape or form did I ever want any
proprietary driver to make use of the framework. As the
project has grown I'm not alone with this sentiment so
lets document that and also fix a few symbols that have
slipped along the way.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>