backports: enable kconfig language on dependencies file
Certain complex features that are backported may be be
limitted to a certain target build configuration. An example
can be if a backported feature is not yet backported with
support for lockdep. In order to avoid build failures with
these types of restrictions allow for specifying build
configuration dependencies on backported upstream kconfig
symbols other than just kernel versioning contstraints.

This adds support for specifying upstream kconfig constaints
other than kernel versioning by adding kconfig language
extensions on the dependencies file. This will update the
copied over upstream Kconfig file for the symbol specified
with the kconfig constraints specified.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>