backports: fix device_release_driver
The new macro, introduced by

commit 9f5876fa9cfb8cd6f450c200043ca7e31915f976
Author: Luis R. Rodriguez <mcgrof@do-not-panic.com>
Date:   Thu Jul 18 16:40:55 2013 -0700

    backports: backport drvdata = NULL core driver fixes

evaluates its argument multiple times, causing crashes as
there are potential side-effects of the called code and/or
the arguments.

Fix this by converting it to static inlines.

Also remove the driver_probe_device() macro that has the
same issue, but tries to wrap a function that isn't even
available outside the driver core code in any kernel.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>