backports: do not copy wrong header files for regulators
Some of the header files for the regulators framework where used by
regulator core, which does not get backported or some of them are also
used by mfd drivers. Now only the headers which are only used by the
regulator driver are copied, it could still be that these headers are
also used in some platform specific code where the driver gets
registered.
This patch then fixed some dependencies which where wrong because of
using the wrong header files.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>
Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>