backports: use sysfs api backport independent from kernel version
In struct class the new sysfs dev_groups member was added with kernel
3.11, but in other structs, like struct bus_type the new member was
added with kernel 3.12. This backport is now needed on different kernel
versions depending on the struct where is was used.
With this patch ATTRIBUTE_GROUPS() will always be used with the new API
and ATTRIBUTE_GROUPS_BACKPORT() will always be used used with the old
API and the patch has to add a KERNEL_VERSION check to add the correct
macro.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>