backports: fix kfifo_put()
kfifo_put() now gets the second parameter by value and not by reference
any more. This replaces the in kernel implementation of kfifo_put()
with a version which is compatible to the in kernel structures but
takes the value and not a reference to the value. We wanted to do this:
 #define kfifo_put(fifo, val) kfifo_put(fifo, &val)

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>