backports: always activate the of.h backports
This removes the KERNEL_HAS_OF_SUPPORT option from backports.
We can not include linux/of.h on kernel < 2.6.34, because this depends
on some sparc and powerpc only headers. The backport code should still
be included in this header file for such kernel versions, because some
recent drivers are using these functions now on all platforms, but they
will use the empty implementations then.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>