backports: move part of multicast handling into an spatch
The netdev_for_each_mc_addr() handling can be done with a
semantic patch and a simple macro encapsulating the addr
dereferencing for the two cases.

In fact, this already fixed a bug in dvb_net.c, although
I think that depends on a newer kernel anyway.

Thanks Julia for the help with iterator name!

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Acked-by: Julia Lawall <julia.lawall@lip6.fr>
Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>