backports: conditionally access net/ieee802154/ with make
This essentially reverts commit cb6f08d3c38c as that caused
issues with copy-list - the scripting when generating the
backport can't prove that net/ieee802154/ isn't needed and
will try to go into it even if it wasn't copied.

Instead, make net/ieee802154/ be entered whenever it might
be needed - which is for CPTCFG_IEEE802154 and CPTCFG_BT.
This is harmless if the directory ends up listed twice but
will cause this problem to disappear.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>